+++ index.js
@@ -1,3 +1,4 @@
 module.exports = {
-  Base: require('./lib/Base')
+  Base: require('./lib/Base'),
+  lit: require('./lib/json2js').lit
 };
