+++ lib/Base.js
@@ -83,7 +83,6 @@ module.exports = generators.Base.extend({
   mergeJson: fileUtils.mergeJson,
   updateJson: fileUtils.updateJson,
   replaceInFile: fileUtils.replaceInFile,
-  copyTemplate: fileUtils.copyTemplate,
-  lit: fileUtils.lit
+  copyTemplate: fileUtils.copyTemplate
 
 });
