+++ lib/file-utils.js
@@ -1,6 +1,6 @@
 const _ = require('lodash');
 const ejs = require('ejs');
-const falafel = require('falafel');
+const json = require('./json2js').json;
 
 exports.mergeJson = function mergeJson(fileName, newContent) {
   const content = this.fs.readJSON(this.destinationPath(fileName), {});
@@ -32,44 +32,10 @@ exports.replaceInFile = function replaceInFile(fileName, regex, templateScope) {
   this.fs.write(this.destinationPath(fileName), newContent);
 };
 
-function Literal(content) {
-  this.content = content;
-}
-
-exports.lit = function lit(string) {
-  return new Literal(string[0]);
-};
-
-function replacer(key, value) {
-  if (value instanceof Literal) {
-    return `lit>>${value.content}<<lit`;
-  } else {
-    return value;
-  }
-}
-
-function json(value) {
-  const js = 'var a = ' + JSON.stringify(value, replacer, 2);
-  const jsFormatted = falafel(js, { ecmaVersion: 6 }, node => {
-    if (node.type === 'Literal') {
-      const lit = /^lit>>(.*)<<lit$/.exec(node.value);
-      if (lit === null) {
-        node.update(`'${node.value}'`);
-      } else {
-        node.update(lit[1]);
-      }
-    }
-    if (node.type === 'Property' && /^[_$a-zA-Z][_$a-zA-Z0-9]*$/.test(node.key.value)) {
-      node.update(`${node.key.value}: ${node.value.source()}`);
-    }
-  });
-  return jsFormatted.toString().replace(/^var a = /, '');
-}
-
-exports.copyTemplate = function copyTemplate(fileName, model) {
+exports.copyTemplate = function copyTemplate(templateFileName, destinationFileName, model) {
   this.fs.copyTpl(
-    this.templatePath(fileName),
-    this.destinationPath(fileName),
+    this.templatePath(templateFileName),
+    this.destinationPath(destinationFileName),
     Object.assign({}, { json }, model)
   );
 };
